+++ CHANGELOG.md
@@ -7,6 +7,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ### Changed
 - Remove `wheel` from `build-system.requires`
 
+### Fixed
+- Fix/update minimal Deployer example in docs
+
 ## 1.0.0 - 2025-03-21
 ### Changed
 - Add more logging to `read_config`